more nt utils into a seperate file
